refactor: move all workflows to workflows/index.ts
[DX-1042]
