--- src/workflows/deposit/depositEth.ts
@@ -1,136 +1,0 @@
-import { Signer } from '@ethersproject/abstract-signer';
-import { TransactionResponse } from '@ethersproject/providers';
-import { DepositsApi, EncodingApi, UsersApi } from '../../api';
-import { parseEther } from '@ethersproject/units';
-import { Core, Core__factory, Registration__factory } from '../../contracts';
-import {
-  getSignableRegistrationOnchain,
-  isRegisteredOnChainWorkflow,
-} from '../registration';
-import { ImmutableXConfiguration, ETHDeposit } from '../../types';
-import { BigNumber } from '@ethersproject/bignumber';
-
-interface ETHTokenData {
-  decimals: number;
-}
-
-async function executeRegisterAndDepositEth(
-  signer: Signer,
-  amount: BigNumber,
-  assetType: string,
-  starkPublicKey: string,
-  vaultId: number,
-  contract: Core,
-  usersApi: UsersApi,
-): Promise<TransactionResponse> {
-  const etherKey = await signer.getAddress();
-
-  const signableResult = await getSignableRegistrationOnchain(
-    etherKey,
-    starkPublicKey,
-    usersApi,
-  );
-
-  const populatedTransaction =
-    await contract.populateTransaction.registerAndDepositEth(
-      etherKey,
-      starkPublicKey,
-      signableResult.operator_signature,
-      assetType,
-      vaultId,
-    );
-
-  return signer.sendTransaction({ ...populatedTransaction, value: amount });
-}
-
-async function executeDepositEth(
-  signer: Signer,
-  amount: BigNumber,
-  assetType: string,
-  starkPublicKey: string,
-  vaultId: number,
-  contract: Core,
-): Promise<TransactionResponse> {
-  const populatedTransaction = await contract.populateTransaction[
-    'deposit(uint256,uint256,uint256)'
-  ](starkPublicKey, assetType, vaultId);
-
-  return signer.sendTransaction({ ...populatedTransaction, value: amount });
-}
-
-export async function depositEthWorkflow(
-  signer: Signer,
-  deposit: ETHDeposit,
-  depositsApi: DepositsApi,
-  usersApi: UsersApi,
-  encodingApi: EncodingApi,
-  config: ImmutableXConfiguration,
-): Promise<TransactionResponse> {
-  const user = await signer.getAddress();
-  const data: ETHTokenData = {
-    decimals: 18,
-  };
-  const amount = parseEther(deposit.amount);
-
-  const getSignableDepositRequest = {
-    user,
-    token: {
-      type: deposit.type,
-      data,
-    },
-    amount: amount.toString(),
-  };
-
-  const signableDepositResult = await depositsApi.getSignableDeposit({
-    getSignableDepositRequest,
-  });
-
-  const encodingResult = await encodingApi.encodeAsset({
-    assetType: 'asset',
-    encodeAssetRequest: {
-      token: {
-        type: deposit.type,
-      },
-    },
-  });
-
-  const assetType = encodingResult.data.asset_type;
-  const starkPublicKey = signableDepositResult.data.stark_key;
-  const vaultId = signableDepositResult.data.vault_id;
-
-  const coreContract = Core__factory.connect(
-    config.l1Configuration.coreContractAddress,
-    signer,
-  );
-
-  const registrationContract = Registration__factory.connect(
-    config.l1Configuration.registrationContractAddress,
-    signer,
-  );
-
-  const isRegistered = await isRegisteredOnChainWorkflow(
-    starkPublicKey,
-    registrationContract,
-  );
-
-  if (!isRegistered) {
-    return executeRegisterAndDepositEth(
-      signer,
-      amount,
-      assetType,
-      starkPublicKey,
-      vaultId,
-      coreContract,
-      usersApi,
-    );
-  } else {
-    return executeDepositEth(
-      signer,
-      amount,
-      assetType,
-      starkPublicKey,
-      vaultId,
-      coreContract,
-    );
-  }
-}
