--- src/workflows/withdrawal/completeERC20Withdrawal.ts
@@ -1,110 +1,0 @@
-import { Signer } from '@ethersproject/abstract-signer';
-import { TransactionResponse } from '@ethersproject/providers';
-import { EncodingApi, UsersApi } from '../../api';
-import {
-  Core,
-  Core__factory,
-  Registration,
-  Registration__factory,
-} from '../../contracts';
-import {
-  ImmutableXConfiguration,
-  ERC20Withdrawal,
-  TokenType,
-} from '../../types';
-import {
-  getSignableRegistrationOnchain,
-  isRegisteredOnChainWorkflow,
-} from '../registration';
-import { getEncodeAssetInfo } from './getEncodeAssetInfo';
-
-async function executeRegisterAndWithdrawERC20(
-  signer: Signer,
-  assetType: string,
-  starkPublicKey: string,
-  contract: Registration,
-  usersApi: UsersApi,
-): Promise<TransactionResponse> {
-  const etherKey = await signer.getAddress();
-
-  const signableResult = await getSignableRegistrationOnchain(
-    etherKey,
-    starkPublicKey,
-    usersApi,
-  );
-
-  const populatedTransaction =
-    await contract.populateTransaction.registerAndWithdraw(
-      etherKey,
-      starkPublicKey,
-      signableResult.operator_signature,
-      assetType,
-    );
-
-  return signer.sendTransaction(populatedTransaction);
-}
-
-async function executeWithdrawERC20(
-  signer: Signer,
-  assetType: string,
-  starkPublicKey: string,
-  contract: Core,
-): Promise<TransactionResponse> {
-  const populatedTransaction = await contract.populateTransaction.withdraw(
-    starkPublicKey,
-    assetType,
-  );
-
-  return signer.sendTransaction(populatedTransaction);
-}
-
-export async function completeERC20WithdrawalWorfklow(
-  signer: Signer,
-  starkPublicKey: string,
-  token: ERC20Withdrawal,
-  encodingApi: EncodingApi,
-  usersApi: UsersApi,
-  config: ImmutableXConfiguration,
-) {
-  const assetType = await getEncodeAssetInfo(
-    'asset',
-    TokenType.ERC20,
-    encodingApi,
-    {
-      token_id: token.data.tokenId,
-      token_address: token.data.tokenAddress,
-    },
-  );
-
-  const coreContract = Core__factory.connect(
-    config.l1Configuration.coreContractAddress,
-    signer,
-  );
-
-  const registrationContract = Registration__factory.connect(
-    config.l1Configuration.registrationContractAddress,
-    signer,
-  );
-
-  const isRegistered = await isRegisteredOnChainWorkflow(
-    starkPublicKey,
-    registrationContract,
-  );
-
-  if (!isRegistered) {
-    return executeRegisterAndWithdrawERC20(
-      signer,
-      assetType.asset_type,
-      starkPublicKey,
-      registrationContract,
-      usersApi,
-    );
-  } else {
-    return executeWithdrawERC20(
-      signer,
-      assetType.asset_type,
-      starkPublicKey,
-      coreContract,
-    );
-  }
-}
