--- src/workflows/withdrawal/index.ts
@@ -1,4 +1,0 @@
-export * from './prepareWithdrawal';
-export * from './completeERC20Withdrawal';
-export * from './completeERC721Withdrawal';
-export * from './completeEthWithdrawal';
